+++ src/main/kotlin/calculator/App.kt
@@ -1,0 +1,7 @@
+package calculator
+
+import calculator.assemble.ObjectAssemble
+
+fun main() {
+    ObjectAssemble.calculatorView().init()
+}
