+++ src/main/kotlin/calculator/application/parser/ParsingProcessor.kt
@@ -1,0 +1,8 @@
+package calculator.application.parser
+
+import calculator.domain.calculator.model.InputOperationCommand
+
+interface ParsingProcessor {
+
+    fun proceed(string: String): List<InputOperationCommand>
+}
