+++ src/main/kotlin/calculator/domain/calculator/Calculator.kt
@@ -1,0 +1,8 @@
+package calculator.domain.calculator
+
+import calculator.domain.calculator.model.InputOperationCommand
+
+interface Calculator<T> {
+
+    fun calculate(inputOperationCommandList: List<InputOperationCommand>): T
+}
