+++ src/main/kotlin/calculator/domain/calculator/model/InputOperationCommand.kt
@@ -1,0 +1,8 @@
+package calculator.domain.calculator.model
+
+import calculator.domain.operation.enums.Operator
+
+data class InputOperationCommand(
+    val number: Int,
+    val nextOperator: Operator
+)
