+++ src/main/kotlin/calculator/domain/operation/command/OperationCommand.kt
@@ -1,0 +1,7 @@
+package calculator.domain.operation.command
+
+import calculator.domain.operation.model.BinaryInputCommand
+
+interface OperationCommand<T> {
+    fun operate(input: BinaryInputCommand): T
+}
