+++ src/main/kotlin/calculator/domain/operation/command/implement/AdditionOperationCommand.kt
@@ -1,0 +1,8 @@
+package calculator.domain.operation.command.implement
+
+import calculator.domain.operation.command.OperationCommand
+import calculator.domain.operation.model.BinaryInputCommand
+
+object AdditionOperationCommand : OperationCommand<Int> {
+    override fun operate(input: BinaryInputCommand) = input.firstNumber + input.secondNumber
+}
