+++ src/main/kotlin/calculator/domain/operation/command/implement/MultiplicationOperationCommand.kt
@@ -1,0 +1,8 @@
+package calculator.domain.operation.command.implement
+
+import calculator.domain.operation.command.OperationCommand
+import calculator.domain.operation.model.BinaryInputCommand
+
+object MultiplicationOperationCommand : OperationCommand<Int> {
+    override fun operate(input: BinaryInputCommand) = input.firstNumber * input.secondNumber
+}
