+++ src/main/kotlin/calculator/domain/operation/model/BinaryInputCommand.kt
@@ -1,0 +1,3 @@
+package calculator.domain.operation.model
+
+data class BinaryInputCommand(val firstNumber: Int, val secondNumber: Int)
