+++ src/main/kotlin/calculator/interfaces/adpater/CalculatorConsoleAdapter.kt
@@ -1,0 +1,15 @@
+package calculator.interfaces.adpater
+
+import calculator.application.port.`in`.CalculatorInputPort
+import calculator.interfaces.input.InputConsole
+import calculator.interfaces.output.OutputConsole
+
+class CalculatorConsoleAdapter(
+    private val calculatorInputPort: CalculatorInputPort<Int>
+) {
+    fun init() {
+        val expression = InputConsole.input()
+        val result = calculatorInputPort.input(expression).toString()
+        OutputConsole.output(result)
+    }
+}
