--- src/main/kotlin/step2/App.kt
@@ -1,7 +1,0 @@
-package step2
-
-import step2.assemble.ObjectAssemble
-
-fun main() {
-    ObjectAssemble.calculatorView().init()
-}
