--- src/main/kotlin/step2/domain/calculator/Calculator.kt
@@ -1,8 +1,0 @@
-package step2.domain.calculator
-
-import step2.domain.calculator.model.InputOperationCommand
-
-interface Calculator<T> {
-
-    fun calculate(inputOperationCommandList: List<InputOperationCommand>): T
-}
