--- src/main/kotlin/step2/domain/operation/command/OperationCommand.kt
@@ -1,7 +1,0 @@
-package step2.domain.operation.command
-
-import step2.domain.operation.model.BinaryInputCommand
-
-interface OperationCommand<T> {
-    fun operate(input: BinaryInputCommand): T
-}
