--- src/main/kotlin/step2/domain/operation/command/implement/AdditionOperationCommand.kt
@@ -1,8 +1,0 @@
-package step2.domain.operation.command.implement
-
-import step2.domain.operation.command.OperationCommand
-import step2.domain.operation.model.BinaryInputCommand
-
-object AdditionOperationCommand : OperationCommand<Int> {
-    override fun operate(input: BinaryInputCommand) = input.firstNumber + input.secondNumber
-}
