--- src/main/kotlin/step2/domain/operation/command/implement/DivisionOperationCommand.kt
@@ -1,8 +1,0 @@
-package step2.domain.operation.command.implement
-
-import step2.domain.operation.command.OperationCommand
-import step2.domain.operation.model.BinaryInputCommand
-
-object DivisionOperationCommand : OperationCommand<Int> {
-    override fun operate(input: BinaryInputCommand) = input.firstNumber / input.secondNumber
-}
