--- src/main/kotlin/step2/domain/operation/command/implement/MultiplicationOperationCommand.kt
@@ -1,8 +1,0 @@
-package step2.domain.operation.command.implement
-
-import step2.domain.operation.command.OperationCommand
-import step2.domain.operation.model.BinaryInputCommand
-
-object MultiplicationOperationCommand : OperationCommand<Int> {
-    override fun operate(input: BinaryInputCommand) = input.firstNumber * input.secondNumber
-}
