--- src/main/kotlin/step2/domain/operation/model/BinaryInputCommand.kt
@@ -1,3 +1,0 @@
-package step2.domain.operation.model
-
-data class BinaryInputCommand(val firstNumber: Int, val secondNumber: Int)
