--- src/main/kotlin/step2/interfaces/adpater/CalculatorConsoleAdapter.kt
@@ -1,15 +1,0 @@
-package step2.interfaces.adpater
-
-import step2.application.port.`in`.CalculatorInputPort
-import step2.interfaces.input.InputConsole
-import step2.interfaces.output.OutputConsole
-
-class CalculatorConsoleAdapter(
-    private val calculatorInputPort: CalculatorInputPort<Int>
-) {
-    fun init() {
-        val expression = InputConsole.input()
-        val result = calculatorInputPort.input(expression).toString()
-        OutputConsole.output(result)
-    }
-}
