--- src/test/kotlin/step2/domain/calculator/StandardCalculatorKoTest.kt
@@ -1,52 +1,0 @@
-package step2.domain.calculator
-
-import io.kotest.core.spec.style.StringSpec
-import io.kotest.matchers.shouldBe
-import step2.domain.calculator.impl.StandardCalculator
-import step2.domain.calculator.model.InputOperationCommand
-import step2.domain.operation.enums.Operator
-
-internal class StandardCalculatorKoTest : StringSpec({
-
-    lateinit var calculator: Calculator<Int>
-    lateinit var defaultInputOperationCommandList: List<InputOperationCommand>
-
-    "덧셈 계산" {
-        defaultInputOperationCommandList = listOf(
-            InputOperationCommand(1, Operator.ADDITION),
-            InputOperationCommand(2, Operator.ADDITION),
-            InputOperationCommand(3, Operator.ADDITION),
-            InputOperationCommand(4, Operator.ILLEGAL_STATE)
-        )
-
-        calculator = StandardCalculator()
-        val result = calculator.calculate(defaultInputOperationCommandList)
-        result shouldBe 10
-    }
-
-    "혼합 계산 - 1" {
-        defaultInputOperationCommandList = listOf(
-            InputOperationCommand(1, Operator.ADDITION),
-            InputOperationCommand(2, Operator.MULTIPLICATION),
-            InputOperationCommand(3, Operator.SUBTRACTION),
-            InputOperationCommand(4, Operator.ILLEGAL_STATE)
-        )
-
-        calculator = StandardCalculator()
-        val result = calculator.calculate(defaultInputOperationCommandList)
-        result shouldBe 5
-    }
-
-    "혼합 계산 - 2" {
-        defaultInputOperationCommandList = listOf(
-            InputOperationCommand(2, Operator.ADDITION),
-            InputOperationCommand(3, Operator.MULTIPLICATION),
-            InputOperationCommand(4, Operator.DIVISION),
-            InputOperationCommand(2, Operator.ILLEGAL_STATE)
-        )
-
-        calculator = StandardCalculator()
-        val result = calculator.calculate(defaultInputOperationCommandList)
-        result shouldBe 10
-    }
-})
